requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.12
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
ibm-cloud-sdk-core==3.23.0
ibm-cos-sdk==2.14.0
ibm-cos-sdk-core==2.14.0
ibm-cos-sdk-s3transfer==2.14.0
ibm-watson==9.0.0
ibm_watsonx_ai==1.3.3
idna==3.10
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.22
langchain-core==0.3.50
langchain-ibm==0.3.10
langchain-text-splitters==0.3.7
langsmith==0.3.24
lomond==0.3.3
numpy==2.2.4
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.2
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tabulate==0.9.0
tenacity==9.1.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
zstandard==0.23.0
